all-msgs

+++ src/main/java/tahub/sdapitahub/controller/TaskController.java
@@ -5,11 +5,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import tahub.sdapitahub.constants.BoardContMsgs;
+import tahub.sdapitahub.constants.TaskMsgs;
 import tahub.sdapitahub.dto.Task.TaskCreateDTO;
 import tahub.sdapitahub.entity.Task;
 import tahub.sdapitahub.service.TaskService;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @RestController
@@ -39,24 +42,39 @@ public class TaskController {
     }
 
     @PostMapping("/task")
-    public ResponseEntity<Task> createTask(@RequestBody TaskCreateDTO taskPostDTO) {
-        Task createdTask = taskService.createTask(taskPostDTO);
-        return ResponseEntity.status(HttpStatus.CREATED).body(createdTask);
+    public ResponseEntity<String> createTask(@RequestBody TaskCreateDTO taskPostDTO) {
+            Task createdTask = taskService.createTask(taskPostDTO);
+            if (createdTask != null) {
+                return ResponseEntity.status(HttpStatus.CREATED).body(TaskMsgs.TASK_CREATED.getMessage());
+            }
+            else{
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(TaskMsgs.ERROR_TASK_CREATE.getMessage());
+
+            }
     }
 
     @PutMapping("/task/{id}")
-    public ResponseEntity<Task> updateTask(@PathVariable Long id, @RequestBody TaskCreateDTO taskPostDTO) {
+    public ResponseEntity<String> updateTask(@PathVariable Long id, @RequestBody TaskCreateDTO taskPostDTO) {
+        try{
         Task updatedTask = taskService.updateTask(id, taskPostDTO);
         if (updatedTask != null) {
-            return new ResponseEntity<>(updatedTask, HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(TaskMsgs.TASK_UPDATED.getMessage(), HttpStatus.OK);
         }
+        }catch (NoSuchElementException e){
+            return new ResponseEntity<>(TaskMsgs.TASK_NOT_FOUND.getMessage(),HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(TaskMsgs.ERROR_TASK_UPDATE.getMessage(),HttpStatus.INTERNAL_SERVER_ERROR);
+
+
     }
 
     @DeleteMapping("/task/{id}")
-    public ResponseEntity<Void> deleteTask(@PathVariable Long id) {
-        taskService.deleteTask(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<String> deleteTask(@PathVariable Long id) {
+        if (taskService.getTaskById(id).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(TaskMsgs.TASK_NOT_FOUND.getMessage());
+        } else {
+            taskService.deleteTask(id);
+            return ResponseEntity.status(HttpStatus.OK).body(TaskMsgs.TASK_DELETED.getMessage());
+        }
     }
 }
